Changed enum mappings on server-client events from integers to strings
This had to be done as for some reason, triggering the event
CLIENT_REQUESTS.STORE_GAME_ATTRIBUTES was causing the creation of 2
players joining the game. The problem is solved by changing the mapping
of the enum to a string value instead o an integer value. Although I am
unsure as to what the root cause is...