Improved cohesion of Lobby component
Prior to this commit, the lobby component conducted many operations that
were better suited to being conducted in the server side. These
operations made use of some information that were already stored in the
server side, but were not actually obtaining them from the server side,
instead keeping its own independent track of these info in the client
side. This commit made it so that these operations and information are
stored in the serverside (such as the messages, user name color mapping)
and are simply passed to the client side react components via socket io
events.

The burden of processing the color selection and message tracking is now
strictly delegated to the serverside and operations located in the
client side that used to conduct these processes have been placed into
the serverside, with improved and simplified logic. This should increase
the cohesion of the Lobby component since it can focus on presentational
code.